server gs 0.65

+++ server/sever_structures.h
@@ -7,8 +7,8 @@
 
 #define MAX_BULLETS 128
 struct  coordinate {
-    int16_t x;
-    int16_t y;
+    float x;
+    float y;
 };
 
 struct bullet {
